+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 <cite>Release contributors</cite>
 
-- 1 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.3.3.2+author%3Amlytvyn+is%3Apr)
+- 2 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.3.3.2+author%3Amlytvyn+is%3Apr)
 
 ### `Cockpit NG` enhancements
 - Inspection: reuse already declared namespace [#1736](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1736)
+- Inspection: move namespace to the root tag [#1737](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1737)
 
 ## [2025.3.3.1]